Sterowanie z obracaniem kamery
Teraz działa poprawnie obracanie kamery wraz ze zmianą movementu

Zmiana minimapy na 5 poziomie

Naprawiony błąd z brakiem kolizji w ukrytym pomieszczeniem w poziomie 5

+++ Cube_Game/Assets/Scripts/CameraMove.cs
@@ -4,18 +4,19 @@ using UnityEngine;
 
 public class CameraMove : MonoBehaviour
 {
+   
     public static CameraMove cameraMove;
     public GameObject targetObject;
     public PlayerMovement playerMovement;
     public GameObject player;
-    public float targetAngle = 0;
-    const float rotationAmount = 1.0f;
+    public int targetAngle = 0;
+    const int rotationAmount = 1;
    float rotationAngle = 90f;
     // Start is called before the first frame update
     void Start()
     {
         playerMovement = GameObject.Find("Player").GetComponent<PlayerMovement>();
-   
+        
     }
 
     // Update is called once per frame
@@ -25,15 +26,51 @@ public class CameraMove : MonoBehaviour
             // Trigger functions if Rotate is requested
             if (Input.GetKeyDown(KeyCode.LeftArrow))
             {
-                targetAngle -= 90.0f;
+                targetAngle -= 90;
+            //if (rotationLocal == 0)
+            //{
+            //    rotationLocal = 90;
+            //}
+            //if (rotationLocal == 90)
+            //{
+            //    rotationLocal = 180;
+            //}
+            //if (rotationLocal == 180)
+            //{
+            //    rotationLocal = 270;
+            //}
+            //if (rotationLocal == 360)
+            //{
+            //    rotationLocal = 0;
+            //}
+            //rotate.transform.localRotation = Quaternion.Euler(0, rotationLocal, 0);
 
-
-            }
+        }
             else if (Input.GetKeyDown(KeyCode.RightArrow))
-            {
-                targetAngle += 90.0f;
-
-            }
+            {           
+            targetAngle += 90;
+            //if (rotationLocal == 360)
+            //{
+            //    rotationLocal = 0;
+            //}
+            //if (rotationLocal == 0)
+            //{
+            //    rotationLocal = 270;
+            //}
+            //if (rotationLocal == 270)
+            //{
+            //    rotationLocal = 180;
+            //}
+            //if (rotationLocal == 180)
+            //{
+            //    rotationLocal = 90;
+            //}
+            //if (rotationLocal == 90)
+            //{
+            //    rotationLocal = 0;
+            //}
+            //rotate.transform.localRotation = Quaternion.Euler(0, rotationLocal, 0);
+        }
 
             if (targetAngle != 0)
             {
@@ -43,6 +80,7 @@ public class CameraMove : MonoBehaviour
         }
 
 
+   
     protected void Rotate()
     {
